I2PSnark: javadocs, source formatting (google format)

+++ apps/i2psnark/java/src/org/klomp/snark/web/FetchAndAdd.java
@@ -5,10 +5,6 @@ package org.klomp.snark.web;
  * with no warranty of any kind, either expressed or implied.
  */
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-
 import net.i2p.I2PAppContext;
 import net.i2p.client.streaming.I2PSocketEepGet;
 import net.i2p.client.streaming.I2PSocketManager;
@@ -25,22 +21,22 @@ import org.klomp.snark.Snark;
 import org.klomp.snark.SnarkManager;
 import org.klomp.snark.Storage;
 
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+
 /**
- *  A cancellable torrent file downloader.
- *  We extend Snark so its status may be easily listed in the
- *  web table without adding a lot of code there.
+ * A cancellable torrent file downloader. We extend Snark so its status may be easily listed in the
+ * web table without adding a lot of code there.
  *
- *  Upon successful download, this Snark will be deleted and
- *  a "real" Snark created.
+ * <p>Upon successful download, this Snark will be deleted and a "real" Snark created.
  *
- *  The methods return values similar to a Snark in magnet mode.
- *  A fake info hash, which is the SHA1 of the URL, is returned
- *  to prevent duplicates.
+ * <p>The methods return values similar to a Snark in magnet mode. A fake info hash, which is the
+ * SHA1 of the URL, is returned to prevent duplicates.
  *
- *  This Snark may be stopped and restarted, although a partially
- *  downloaded file is discarded.
+ * <p>This Snark may be stopped and restarted, although a partially downloaded file is discarded.
  *
- *  @since 0.9.1 Moved from I2PSnarkUtil
+ * @since 0.9.1 Moved from I2PSnarkUtil
  */
 public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnable {
 
@@ -61,19 +57,17 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
     private Thread _thread;
     private EepGet _eepGet;
 
-//    private static final int RETRIES = 3;
+    //    private static final int RETRIES = 3;
     private static final int RETRIES = 20;
 
     /**
-     *   Caller should call _mgr.addDownloader(this), which
-     *   will start things off.
+     * Caller should call _mgr.addDownloader(this), which will start things off.
      *
-     *   @param dataDir null to default to snark data directory
+     * @param dataDir null to default to snark data directory
      */
     public FetchAndAdd(I2PAppContext ctx, SnarkManager mgr, String url, File dataDir) {
         // magnet constructor
-        super(mgr.util(), "Torrent download",
-              null, null, null, null, null, null);
+        super(mgr.util(), "Torrent download", null, null, null, null, null, null);
         _ctx = ctx;
         _log = ctx.logManager().getLog(FetchAndAdd.class);
         _mgr = mgr;
@@ -83,58 +77,56 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
         byte[] fake = null;
         try {
             fake = SHA1.getInstance().digest(url.getBytes("ISO-8859-1"));
-        } catch (IOException ioe) {}
+        } catch (IOException ioe) {
+        }
         _fakeHash = fake;
     }
 
-    /**
-     *  Set off by startTorrent()
-     */
+    /** Set off by startTorrent() */
     public void run() {
         _mgr.addMessageNoEscape(_t("Requesting torrent file: {0}", urlify(_url)));
         File file = get();
-        if (!_isRunning)  // stopped?
-            return;
+        if (!_isRunning) // stopped?
+        return;
         _isRunning = false;
         if (file != null && file.exists() && file.length() > 0) {
             // remove this in snarks
             _mgr.deleteMagnet(this);
             add(file);
         } else {
-            _mgr.addMessageNoEscape(_t("Failed to retrieve torrent file: {0}", urlify(_url)) +
-                            ((_failCause != null) ? (": " + DataHelper.stripHTML(_failCause)) : ""));
+            _mgr.addMessageNoEscape(
+                    _t("Failed to retrieve torrent file: {0}", urlify(_url))
+                            + ((_failCause != null)
+                                    ? (": " + DataHelper.stripHTML(_failCause))
+                                    : ""));
         }
-        if (file != null)
-            file.delete();
+        if (file != null) file.delete();
     }
 
     /**
-     *  Copied from I2PSnarkUtil so we may add ourselves as a status listener
-     *  @return null on failure
+     * Copied from I2PSnarkUtil so we may add ourselves as a status listener
+     *
+     * @return null on failure
      */
     private File get() {
-        if (_log.shouldDebug())
-            _log.debug("Requesting [" + _url + "]");
+        if (_log.shouldDebug()) _log.debug("Requesting [" + _url + "]");
         File out = null;
         try {
             out = SecureFile.createTempFile("torrentFile", null, _mgr.util().getTempDir());
         } catch (IOException ioe) {
             _log.error("Temporary storage error", ioe);
             _mgr.addMessage("Problem writing file to temp directory: " + ioe.getMessage());
-            if (out != null)
-                out.delete();
+            if (out != null) out.delete();
             return null;
         }
         out.deleteOnExit();
 
         if (!_mgr.util().connected()) {
             _mgr.addMessage(_t("Opening the I2P tunnel") + "...");
-            if (!_mgr.util().connect())
-                return null;
+            if (!_mgr.util().connect()) return null;
         }
         I2PSocketManager manager = _mgr.util().getSocketManager();
-        if (manager == null)
-            return null;
+        if (manager == null) return null;
         _eepGet = new I2PSocketEepGet(_ctx, manager, RETRIES, out.getAbsolutePath(), _url);
         _eepGet.addStatusListener(this);
         _eepGet.addHeader("User-Agent", I2PSnarkUtil.EEPGET_USER_AGENT);
@@ -143,16 +135,15 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
                 _log.info("Transfer successful: " + _url + " (Size:" + out.length() + " bytes)");
             return out;
         } else {
-            if (_log.shouldInfo())
-                _log.info("Transfer failed: " + _url);
+            if (_log.shouldInfo()) _log.info("Transfer failed: " + _url);
             out.delete();
             return null;
         }
     }
 
     /**
-     *  Tell SnarkManager to copy the torrent file over and add it to the Snarks list.
-     *  This Snark may then be deleted.
+     * Tell SnarkManager to copy the torrent file over and add it to the Snarks list. This Snark may
+     * then be deleted.
      */
     private void add(File file) {
         _mgr.addMessageNoEscape(_t("Torrent downloaded from {0}", urlify(_url)));
@@ -161,10 +152,16 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
             in = new FileInputStream(file);
             byte[] fileInfoHash = new byte[20];
             String name = MetaInfo.getNameAndInfoHash(in, fileInfoHash);
-            try { in.close(); } catch (IOException ioe) {}
+            try {
+                in.close();
+            } catch (IOException ioe) {
+            }
             Snark snark = _mgr.getTorrentByInfoHash(fileInfoHash);
             if (snark != null) {
-                _mgr.addMessage(_t("Torrent with this info hash is already running: {0}", snark.getBaseName()));
+                _mgr.addMessage(
+                        _t(
+                                "Torrent with this info hash is already running: {0}",
+                                snark.getBaseName()));
                 return;
             }
 
@@ -177,25 +174,30 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
             if (torrentFile.exists()) {
                 if (_mgr.getTorrent(canonical) != null)
                     _mgr.addMessage(_t("Torrent already running: {0}", name));
-                else
-                    _mgr.addMessage(_t("Torrent already in the queue: {0}", name));
+                else _mgr.addMessage(_t("Torrent already in the queue: {0}", name));
             } else {
                 // This may take a LONG time to create the storage.
                 boolean ok = _mgr.copyAndAddTorrent(file, canonical, _dataDir);
-                if (!ok)
-                    throw new IOException("Unknown error - check logs");
+                if (!ok) throw new IOException("Unknown error - check logs");
                 snark = _mgr.getTorrentByBaseName(originalName);
-                if (snark != null)
-                    snark.startTorrent();
-                else
-                    throw new IOException("Unknown error - check logs");
+                if (snark != null) snark.startTorrent();
+                else throw new IOException("Unknown error - check logs");
             }
         } catch (IOException ioe) {
-            _mgr.addMessageNoEscape(_t("Torrent at {0} was not valid", urlify(_url)) + ": " + DataHelper.stripHTML(ioe.getMessage()));
+            _mgr.addMessageNoEscape(
+                    _t("Torrent at {0} was not valid", urlify(_url))
+                            + ": "
+                            + DataHelper.stripHTML(ioe.getMessage()));
         } catch (OutOfMemoryError oom) {
-            _mgr.addMessageNoEscape(_t("ERROR - Out of memory, cannot create torrent from {0}", urlify(_url)) + ": " + DataHelper.stripHTML(oom.getMessage()));
+            _mgr.addMessageNoEscape(
+                    _t("ERROR - Out of memory, cannot create torrent from {0}", urlify(_url))
+                            + ": "
+                            + DataHelper.stripHTML(oom.getMessage()));
         } finally {
-            try { if (in != null) in.close(); } catch (IOException ioe) {}
+            try {
+                if (in != null) in.close();
+            } catch (IOException ioe) {
+            }
         }
     }
 
@@ -203,14 +205,13 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
 
     @Override
     public synchronized void startTorrent() {
-        if (_isRunning)
-            return;
-         // reset counters in case starting a second time
-         _remaining = -1;
-         // leave the total if we knew it before
-         //_total = -1;
-         _transferred = 0;
-         _failCause = null;
+        if (_isRunning) return;
+        // reset counters in case starting a second time
+        _remaining = -1;
+        // leave the total if we knew it before
+        // _total = -1;
+        _transferred = 0;
+        _failCause = null;
         _started = _util.getContext().clock().now();
         _isRunning = true;
         _active = false;
@@ -221,8 +222,7 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
     @Override
     public synchronized void stopTorrent() {
         if (_thread != null && _isRunning) {
-            if (_eepGet != null)
-                _eepGet.stopFetching();
+            if (_eepGet != null) _eepGet.stopFetching();
             _thread.interrupt();
         }
         _isRunning = false;
@@ -250,7 +250,7 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
     }
 
     /**
-     *  @return torrent file size or -1
+     * @return torrent file size or -1
      */
     @Override
     public long getTotalLength() {
@@ -258,7 +258,7 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
     }
 
     /**
-     *  @return -1 when done so the web will list us as "complete" instead of "seeding"
+     * @return -1 when done so the web will list us as "complete" instead of "seeding"
      */
     @Override
     public long getRemainingLength() {
@@ -267,7 +267,7 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
     }
 
     /**
-     *  @return torrent file bytes remaining or -1
+     * @return torrent file bytes remaining or -1
      */
     @Override
     public long getNeededLength() {
@@ -280,8 +280,7 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
             long time = _ctx.clock().now() - _started;
             if (time > 1000) {
                 long rv = (_transferred * 1000) / time;
-                if (rv >= 100)
-                    return rv;
+                if (rv >= 100) return rv;
             }
         }
         return 0;
@@ -306,17 +305,27 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
 
     // EepGet status listeners to maintain the state for the web page
 
-    public void attemptFailed(String url, long bytesTransferred, long bytesRemaining, int currentAttempt, int numRetries, Exception cause) {
+    public void attemptFailed(
+            String url,
+            long bytesTransferred,
+            long bytesRemaining,
+            int currentAttempt,
+            int numRetries,
+            Exception cause) {
         if (bytesRemaining >= 0) {
             _remaining = bytesRemaining;
         }
         _transferred = bytesTransferred;
-        if (cause != null)
-            _failCause = cause.toString();
+        if (cause != null) _failCause = cause.toString();
         _active = false;
     }
 
-    public void bytesTransferred(long alreadyTransferred, int currentWrite, long bytesTransferred, long bytesRemaining, String url) {
+    public void bytesTransferred(
+            long alreadyTransferred,
+            int currentWrite,
+            long bytesTransferred,
+            long bytesRemaining,
+            String url) {
         if (bytesRemaining >= 0) {
             _remaining = bytesRemaining;
             _total = bytesRemaining + currentWrite + alreadyTransferred;
@@ -325,7 +334,13 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
         _active = true;
     }
 
-    public void transferComplete(long alreadyTransferred, long bytesTransferred, long bytesRemaining, String url, String outputFile, boolean notModified) {
+    public void transferComplete(
+            long alreadyTransferred,
+            long bytesTransferred,
+            long bytesRemaining,
+            String url,
+            String outputFile,
+            boolean notModified) {
         if (bytesRemaining >= 0) {
             _remaining = bytesRemaining;
             _total = bytesRemaining + alreadyTransferred;
@@ -334,7 +349,8 @@ public class FetchAndAdd extends Snark implements EepGet.StatusListener, Runnabl
         _active = false;
     }
 
-    public void transferFailed(String url, long bytesTransferred, long bytesRemaining, int currentAttempt) {
+    public void transferFailed(
+            String url, long bytesTransferred, long bytesRemaining, int currentAttempt) {
         if (bytesRemaining >= 0) {
             _remaining = bytesRemaining;
         }
